backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
# Pin para evitar problemas TLS/compatibilidad en Python 3.13 con Mongo Atlas
motor==3.3.2
pymongo==4.6.3
dnspython>=2.6.1
certifi>=2023.7.22
python-jose[cryptography]>=3.3.0
# passlib 1.7.4 + bcrypt 3.2.2: combinación estable (evita warnings/roturas con bcrypt 4.x)
passlib==1.7.4
bcrypt==3.2.2
python-multipart>=0.0.9
python-dotenv>=1.0.0
pyswisseph>=2.10.0
pytz>=2023.3
stripe>=7.0.0
timezonefinder>=6.1.0
requests>=2.31.0
Pillow>=10.0.0
google-generativeai>=0.3.0
